Visual changes + Bug fixes to Raid
Item bubbles fly to the heart when an item is picked up, and the heart gets darker based on percentage of max loot weight.
The Lungs start to flash red when time is starting to end.
The Level can be stopped by tapping the lungs.
Fixes to audio playing while the level has ended etc.