Added Basic Furniture handling for SoulHome Rooms
Now the rooms can actually have Furniture in them and they also show in the room.
This means the rooms now have slots for the furniture to populate and there is a check to see if the slot is already taken.

Currently the furnitures placeholders are randomly placed. This changes when the servers gets plugged into the soulrooms.

Also refactored the RoomData a little bit.