Furnitures can now be moved inside SoulHome rooms.
Added a way to move to furnitures inside soulhome, icluding assining them to new "slot" on the floor and clear the old slots for later use.
The furniture are opaque while moving to show that they are selected and in process of moving.
Additionally the furniture are now showed in their new position provided that the slot(s) they are hovering over is/are free. If the object isn't over a valid slot then the furniture hover at the location of players touch/click.
If the slot isn't valid when released the object returns to original position.

Refactored the click/touch detection, that now allows for holding down on the screen to move objects as well as allows the player to cancel their action by moving their finger/cursor away from the object before releasing.

Made the rooms 25% higher and made the camera zoom in little less when going into a room.

Fixed bug that caused the soulhome aspectRatio to bug out when backing out of and then returning into the soulHome again.
Slowed the scrolling down a lot while using touchcontrols.