Added methods to add and remove furniture from rooms.
Furniture can now be added to the rooms and be taken away from there.
Added an FurnitureTray to show and select the items in storage (not properly connected yet).
Furniture's are not yet removed or added to the tray when added or removed to the rooms. (Another thing that realistically needs the storage to be connected in order to differentiate between items.)