Fixed multiple bugs from SoulHome
The Camera should now properly place to the bottom of the soulhome when opened for the first time.
The soulhome bar is now set to the middle of the view instead of trying to clamp it if the view is larger than the clamping area.
Fixed Soulhome view scrolling in WebGL.
Fixed Furniture movement not working properly in WebGL.
A room will no longer be selected while the scrolling is in movement.
(Very slight movement is allowed.)