Improved the furniture moving in SoulHome
Furniture movements are no longer automatically saved when they are set down, instead they need to be confirmed by pressing the save changes button.
Additionally any unsaved changes can be reverted by pressing the discard changes button.
Unsaved changes are also discarded if the player leaves SoulHome or closes the editing mode. (More on the below.)
Added an editing mode during which the furniture are moved.
While editing mode is on the one second delay to pick furniture up is waived.
Furniture can now be moved while out side off a specific room and can be moved to any room.
Player can now directly move to the room above or below the one they currently are in.
(Mobile) While in SoulHome the player can use landscape mode of the phone to use and browse SoulHome. (UI has not been adjusted completely yet for landscape.)

Fixed a bug that caused the player to leave the room they are in when sliding the finger to outside of the room.
Fixed a bug that caused the view to get stuck when leaving a room.