Quick fixes to SoulHome
Added more checks to determine the correct zoom level in accordance to the used platform.
Changed the method on checking for doubletaps as the old methods didn't work properly on android.
The soulhome no longer tries to set Screenrotation properties if the platform doesn't suppor them.
Furniture can no longer be picked up in towerview if the editing mode isn't active.