Added handling to the FurnitureTray
Furniture can now be properly added to the FurnitureTray and taken away from it. (ie. new furniture elements are added to the tray when moved there from the soulhome and are removed when moved into the soulhome)
This is also able to handle saving the changes or reverting them.

Added Furniture names into the FurnitureTray slots.

The scroll speed on touchdevices now adjust automatically to the screensize. (Not fully perfect but very close)

Fixed few bug with furniture placement handling between the soulhome and tray.