Popup windows added to SoulHome
SoulHome now checks if there are unsaved changes when the player tries to leave SoulHome or tries to close editing mode.
If there are, a popup shows informing the player of the fact and prompts the player to choose if they actually want to proceed (and revert unsaved changes)
Additionally, there is a popup that shows up when the player saves or reverts the changes in SoulHome, respectively.