Cleaned the MainScreenController script a little bit.
Created a handler to check if the player is click or touching the screen that returns the state of the click.
MainScreenController now used this script to handle clickstate instead of
doing it itself. ( This shortens the code inside MainScreenController by quite a bit.)

Fixed a bug with scrolling didn't stop with mouse controls (mainly affects WebGL)