Soulhome can now create FurnitureSlots for rooms dynamically.
You can assign values to tell how many rows and columns of FurnitureSlots are desired and the Room will not create that many slots upon initialization.

No slots are set in the prefab itself anymore (ie. all of the slots are now created by the code.)

Size of the slots depends on the FloorSize (that needs to be assigned manually) and the number of slots.

Updated the Furniture position setting code to handle variable slot sizes.

Updated Furniture Scaling to be a less rigid and depend on the slotgrid size.

Furniture are now set to the bottom of the slot instead of the middle.

FurnitureSlotGrid is now 20x6 instead of 8x3.

Slightly tweaked the furniture sizing to fit the now slot sizes.

Updated the SoulHome text fonts.

Changed the IntroVideo skip to use ClickStateHandler.