Redid how moving the furniture is handled in SoulHome.
Furniture are no longer automatically placed down when released and instead stay in moving mode until the place down button is pressed or the player starts to interact with another furniture at which point the previous furniture that wasn't set down is reset.
Furniture can now be rotated provided that they have rotated states set (ie. sprite for the rotated form is set)
Changed the soulhome UI layout slightly. (Made the soulhome view itself slightly smaller, while gave more space to Furniture tray and the bottom buttons.)
Added a button to exit the current room in soulhome.
Removed the doubletap exit while, editing mode is on and reduced the grace period for detecting the doubletap.
Added The first proper furniture (Taakka_Sofa) and added it into the furniture tray.
This furniture is also the one that is able to be rotated.
Added new list to FurnitureReference that should be the one to hold all of the furniture prefabs stored. (And same also for FurnitureTrayRefrence)
Added buttons that follow the selected furniture that player is interacting with.
Updated ResizeCollider script to take collider offset into account and made it possible to call it when needed.
When the edit mode is closed through the warning popup it now properly deselects the current furniture. (Same for closing the soulhome)
SoulHome now sets the camera bounds properly in accordance to the number of rooms in soulhome.
Removed references to the old input system. (Both soulhome and my ClickState wrapper)
Cleaned some of the older code and moved it into their own methods.
Cleaned the soulhome prefab folder.
Fixed some older bugs related to incorrect furnitureslot calculations that allowed for values that were higher than the number of slots.