Added more furniture to SoulHome.
Added Mirror_Taakka and Floorlamp_Taakka to SoulHome
Updated the sprite of Sofa_Taakka.
Updated the code on how to get furniture size and made it also give the rotated size if the furniture is rotated.
Added the new furniture to the Furniture Tray.
Added a hopefully temporary workaround fact that the placeholder and actual furniture have their size set differently.