Fixed some bugs with furniture position detection.
Furniture position check tried to check the position on the very edge of the slot when setting the furniture down properly. This caused problems presumably due to rounding errors that caused the check to sometimes see the wrong slot. Now the check is made on the middle of the slot instead to avoid this.
The furniture position didn't check the slot width properly until it was being set down. This resulted in the furniture position to be slightly off on the x-axis while the furniture was in hovering mode. This is now fixed.