Improvements to FurnitureTray and Furniture ControlButtons
FurnitureTray no longer gets locked until the furniture is actually pulled into the Soulhome.
The furniture no longer hovers where you touch if it already is a part of the tray.
Changed the furniture control buttons to be a static part of the UI next to the Tray instead of hovering by the furniture.
Additionally made the tray slightly shorter to make space for the buttons.
Added a method to unselect the temporary furniture selection.

Fixed a few bugs that caused the furniture to not reappear in the tray immidiately when the location it was placed wasn't valid.
Fixed a bug that caused the furniture slots to not be freed correctly.