Multiple fixes to FurnitureTray and Furniture placing.
Mostly reverted the change to FurnitureSlot checking that I made earlier, now it just checks veeeery slightly higher than it originally did to avoid the rounding error. (The change caused a lot of unintended errors when the furniture was not above a valid slot.)
The furniture now gets removed correctly if it doesn't have a valid position saved (ie. it was never properly set down yet) when trying to place into invalid location.
Removed a redundant check of whether the player touched the soulhome screen area that deselected the furniture if they didn't as this caused unintended deselections.
FurnitureTray item is now unselected if the hidden tray item is revealed as this should mean that the selection is no longer needed.