Added checks for furniture that are supposed to be next to walls.
Added Sink_Schrodinger.
Added values to furniture that determine if they are on the floor, on the  wall or on the ceiling.
Added checks that make sure that only furniture that are supposed to be next to wall can only be placed there.
These furniture are also automatically rotated according to which wall they are next to.
Fixed a bug that caused music to not loop properly.
Fixed a bug that caused confirm popup sounds to not play properly.
Renamed Towers camera to Tower Camera to represent what it properly is.
Default font has been changed from LiberationSans to WorkSans.