Updated soulhome sprites and added basic "audiomanager"
Updated Taakka set furniture to the newest sprites.
Added a rudimentary AudioManager to SoulHome that should contain all the music and sounds that soulhome needs.
The sounds can then be played from the manager instead of individual buttons separately.
Also allows sounds to play even if the object involved is disabled (as this would prevent playing the sound if the sound was attached to the object itself.)