Updated the Soulhome room layout
Changed the Soulhome room to be a new 8x20 sized room.
The FurnitureSlots should now better align and scale with the floor and the depth.
When the room is now placed into the Soulhome tower its position is now set dynamically.
Fixed a bug that caused Furniture to check wrong variable while freeing FurnitureSlots.
Fixed the size and collidersize of some furniture.