Fixes to Avatar PathNavigation
Fixed an error that caused out of index error, when checking the direction that the avatar needs to head towards after hitting a furniture.
Reduced the amount of unnecessary collision checks at the destination.
Fixed Avatar getting the wrong furniture corner on the x axis.
Also added a check to remove an error when opening SoulHome while GameAnalytics is not initialized. (Mainly affects editor.)