Remade large parts of the FurnitureTray handling
Created a proper list of furniture that the clan has in their storage and soulhome.
This list contains separate objects for each type of furniture and each individual furniture are then contained in a list inside these objects.
The objects expose the name of the objects and the number of furniture contained within for easy readonly access.
Furnitures of same type are now in tray only once.
Added a counter to show how many furniture are available in the tray.
Now the object on the tray is only hidden if there are no more available furniture of that type. (This can be later changed to just graying it out)
Reverting the changes updates the amount of available furniture automatically.
No longer uses the actual furnitureTrayObject when taking furniture form the tray, but instead creates a copy of it.

Fixed a bug that caused furniture resetting method to error out if the furniture was placed back to the tray.
Fixed a bug that caused Avatar pathtracing to select a wrong position instead of the next corner, if the new position was behind the furniture when hitting the furniture from the side.
Fixed a bug that caused popups in soulhome to not work.