Cleaned the main clan page a little bit.
Switched the main clan page to use the UIOverlay instead of using one that is built-in directly into page.
Fixed the missing images in the clan page. (These are placeholders until proper images are created, but it should work for now.)
Switched the clan listing page to use the new layout.
Changed the clan button on the homepage to point back towards the clan page if the player is a member of a clan. (Still directs you toward the clan creation/join page if not.)
Created a script to handle the above properly.
Changed the back button in clan listing/join page to also direct you back out if it isn't also part of the clan creation page.