Redesigned lot of the character class files for future use.
Created Enums for each character and reserved a lot (~255 per class) for future use.
Also changed the class enums to new names and bit shifted them 8 bits to the left.
These are then combined to determine an unique id for each character.
CustomCharacter is now responsible for each character and CharacterClass handles Class information.
BattleCharacter the combines these two and gets sent to the battle scene.
Added a new CharacterArray for selected characters to PlayerData.
Streamlined a little how the character id's are sent from one script to the next.
Redesigned how the GameSettings script handles the battle character prefabs and made it compatable with the new ids.

Co-Authored-By: Niklas <129056239+Niklas603619@users.noreply.github.com>