Updated CharacterGallery character generation
Now no longer uses static preset character slots in CharacterGallery but instead now instantiates them when gallery is initialized.
Created a reference sheet for the sprites and names used in gallery characters that are put in place during initialization.
Redesigned the characterslot prefabs and the associated buttons to allow the slot to retain the sprite and name of the character even when that character is currently in use.
DraggableGallery script now also stores the CharacterId of the character.
Various other miscellaneous changes that should help placing information into right places in the future.

The CharacterGallery is currently a little bit of a mess as a result of these changes but Markus should hopefully fix them asap.