Created a WindowNavigation script to handle window changes
If effect this is the same as the functionality of the NaviButton script, but without the need for a button.
This allows an easy way to change windows from code to the window provided to the script.
NaviButton now extends the WindowNavigation as there is no need to repeat the code unnecessarily.