Updated the way that Login is handled when starting the app
The game now tries to login the player before they reach the mainmenu window.
If the ServerManager fails to login the player with the token, then it prompts the player to login or register.
Restructured the loader scene to move between each section of the load more fluently.
A lot of the code that was previously in a single script are now distributed over multiple.
Since the player now has to login to start playing the check for character intro now checks the PlayerData and whether a character is selected.
Proper loading screens need to be made later, but this should suffice for now.
Also fixed a bug that caused the Local Model script to fail to find the correct PlayerData object due to the unique identifier being saved in a weird way.
Additionally, the signin screen now return the player to the mainmenu on successful login if the login was called from Main Menu.