Storage and background data now uses actual furniture data.
Storage works again.
The ClanFurniture model is now updated to reflect the actual data that is desired there.
Same for GameFurniture script.
ClanFurniture contains the actual furniture objects that the clan owns (ID, position and room of soulhome if applicable and rotation status).
GameFurniture contains the static information about that furniture (size, where the furniture is placed, weight, value).
A default ClanFurniture set is now put into place when failing to get clans furniture from the server.
The GameFurniture contained in DefaultModel is updated.
Also disabled RaidRoom initilization cause it is very hardcoded to the old set and has no actual errorchecking in it.
Created an asset to serve as a StorageFurnitureReference sheet to contain the sprites and names of the furniture in storage.
The storage now uses this sheet to get the images and names to show in it.