The new battlepopup is now functional and in use.
Disabled the old lobby panel as it is no longer needed.
The other game modes than Custom Game are currently not available yet as they are yet to be done.
The scrollable room window is not yet done. but it should eventually allow for scroll through all of the available rooms with a full (or close to it) screen.
Creating a new room and joining an existing room does currently work.
The version number, ping and the number of online players is now visible from the top of the popup panel.
Made some changes to the SwipeUI script and how it handles the fact that there is now one less panel on the main screen.
Swipe UI is now able to take a variable amount of panels rather than a static 5.