Redesigned the way Custom- and BattleCharacters are stored.
The CustomCharacter and BattleCharacter lists in PlayerData are now actually modifiable but private.
This means that the CustomCharacters and BattleCharacters in these are now updated properly.
The getters for these now sends a readonly version out of these lists.
Added a method to update the contents of the lists in PlayerData.
This mainly applies to BattleCharacters.
This might need to be updated later to reduce the load caused by this update, as it effectively rebuilds the BattleCharacter list.
Added a method to fetch the CharacterClasses from LocalModels.
Added a method to properly get the BattleCharacter lists that only contain the selected characters.
Added a method to fill the formentioned list with empty dummy objects to pad out the list.
Now PhotonLobby actually also receives the BattleCharacters according to the selected characters.