Sorting now shows the sorted stat in the quick preview
The type of stat showed changes when the "sorted by" changes to match.
Added a possibility to sort by furniture value.
If the "sorted by" is anything other than name then the name is set above the image.
Re-added material field to the info screen.
Storage refresh button is now on the top bar instead of the top corner of the storage area.
Removed some of the redundant buttons from the top bar.