Added a "safe" method to reference sheet reading
This makes sure that you get the object that you are looking for regardless of the order that the objects are on the list. This just is slower to process, but it shouldn't matter that much.
Renamed the old method as "fast" method.
Changed the Avatar reference script to use the "safe" method to fix some of the issues on that side.