Fixed various bugs with the previous button swap
Fixed Store button not working when not in main menu
Fixed Dailytask button sending player to the old independent window instead of MainMenu when not in main menu already.
Fixed the size of the button text in overlay.
Fixed window derping out when getting to the main menu for the first time.