The game now actually uses the BaseCharacter classes
The game now initializes all of the available Characters into a list in the LocalModels object.
This replaces the old CharacterClass object and it's list.
This new BaseCharacter list is then used to create the CustomCharacter list.
CustomCharacter is now given the values in the BaseCharacter.
Also The CustomCharacter now stores the BaseCharacter Object for quick future referencing.