Consolidated all of the furniture reference scripts into one
This means that all of the objects and image are now placed into a singular reference sheet asset.
This should make it easier to maintain the furniture that are used in the game since they no longer need to be placed into three different places.
Also cleaned some of the wacky GetChild calls form some of the scripts by just giving the script the needed object directly.