First part of debugger update.
Fixed most of the messed up object positioning. (No more buttons outside of screen.)
Full message panel now contains three separate fields for Client number and Timestamp, the main body of the message and the trace the the message.
Main body is highlighted with the colour of the message type.
Messagepanel no longer closes when trying to open a new message.
In file opener user can now choose the client/logbox they want to give the logfile to.
Multiple logfiles can now be chosen before trying to send the m to the parser and thus open them.
LogBoxHandler now send the selcted massage directly to the messagepanel without  having to use LogBoxController in between.