Timeline blocks now show the information separately for each client
The timeline filtering now takes this into account.
The timestamp of the block you are hovering over is now displayed on top of the screen.
Clicking the timestamp block now tries to take you to its position on the logbox or the closest available message.
This is a little derpy still though.
Also adjusted the text size of the logbox.