Added filter syncing to debug.
If filter syncing is on on the timeline or logbox it tries to apply the same filter change to other logboxes or timeline with sync on.
If sync is off things are not synced for that component.
Filterstates from other filters are not automatically set when sync is enabled.
Also moved LogFileOpeningPanel code up a parent to stop it from being disabled when panel is closed.