Updated AudioManager to make adding new AudioSources easier
One can now add new AudioSources by pressing a button on AudioManager script while in editor.
You can and should include a name for these new objects directly form the inputfield on the script.
You can also define the AudioType for the sound (this is still a little incomplete)
Adding the Audiosource also automatically adds it into the audiolist.
Also turned AudioManager into a prefab so it could be used outside of SoulHome.
Sounds can now be searched with the audio name, type or id.
Type is currently tied to enums and you can properly add more of them easily yet.
ID is the position of the sound in the audiolist, with 0 being reverved to empty and 1 being reserved to Music.