Second part of the Photon update
The client now properly services the Photon after joining a lobby.
Creating and joining a room is now possible again.
Updated InRoom panel to use the updated photon scripts.
Disabled Unique nickname setting as it caused a lot of errors.
The Battle PopUp root object is no longer the toggleable one (the MainPanel child object is now the one that is toggled to show and hide it).
The previous change now also means that the main battle popup code is now always active while in main menu and places the player back into the lobby automatically.
The BattlePopup can now be closed by pressing the battle button again.