Added a generic Dictionary to store information into DataCarrier
Things can be added into this storage with AddData method.
You need to provide the Datatype and an empty key to store the data.
To get this data back one should use GetData method.
Similarly to above you need to provide the datatype to the data stored and the key it is stored in.
Also by default the Get method will also clear the key field from the Dictionary.
This can be avoided by providing the Get- method an optional false boolean as a second parameter.