Updated the Sprite ordering in SoulHome
Room Object itself still uses 0.
WallPaper uses 1.
Wall furniture currently uses 11. With the rest of the 1X being reserved for them.
FloorNonblock uses 20-99 with the items in the back of the room using 2X and the front of the room using 9X.
Floor furniture use 100-899 with same logic as above. (I could change this to only be 100-189.) Specifically X03.
Avatar uses X06 to make sure it doesn't overlap with over object causing sprite clipping.
900-999 are reserved for potential special effect that can come over the room.