Rooms are now also on different layers from one another.
Rooms use multiples of 1000 to determine their drawing order.
This is made in order to include the ceiling grid eventually as part of the ceiling furniture would be behind the higher floors floor.
Also allows us to to see the top of the ladder coming from other floors.
Also updated the Roomsprite.
Added the ladder sprites and an alternative floor sprite with a hole for the ladder.