Added a base for the BaseWindow structure.
Uses the same scaler as the overlay to determine the size of the content area.
The content area is divided to Header, tabline and the proper content area.
Header and tabline are handed a fixed size with content taking the area that is remaining.
Header and tabline can be disabled at which case the area is given to content.