Switched data delivery to battle from Realtime to Quantum
Created a BattleCharacterBase class into Quantum to store the data sent there.
Also stored the player position in RunTimePlayer as it is needed to facilitate correct spawning.
There is still a bunch of old code that is being sent to Realtime that should be culled, but some probably should stay there to facilitate proper tracking of the status players in the lobby room.