SwipeUI bug fix
- Switched Main Menu's CurrentPage to be set directly as 2 in Awake()

When returning from intro story to the main menu, the page number would be set as the page number that was last used in the intro story. This would cause errors when the page number was more than the max page number in main menu. Current solution should bring the player back to home screen after choosing a character in the intro story.