Updated Dailytask fetching servercall and local data structure
Updated the GetPlayerTasksFromServer to new server endpoint.
Updated the format of ServerPlayerTask and PlayerTask to match the new changes to server.
ServerPlayerTasks and PlayerTasks are no longer used.
Moved PlayerTask to no longer be a nested object inside PlayerTasks.
Player tasks are now stored inside a List of PlayerTask objects instead of PlayerTasks object.
Fixed all of the references that were broken by this.