Added BattleStart and BattleStory Windows
BattleStart is shown after the Battle start command is received.
There is 5 second delay before moving to battle in proper.
The 5 seconds start from when the start message is sent, to sync the battle start between players. (Start time is sent with the message)
The BattleStory window is shown when battle is closed.
There currently is no functionality in this window except closing and returning to main menu.