Created Route and RouteSection Objects
Recreated the current routes using these.
Route has a start point and a end point.
Also has default speed that is used if no other speed is set for the sections.
If no sections are set the route does directly from start to the end with the set default speed.
If sections are set then the route travels through the points offered in the section.
The section only has the end point of the section since the start is either the endpoint of the previous section or the start point of the route.
AnimationCurve is used to determined the exact path the route takes to the next section end point.
Also the speed can be separately determined for each section to use.
If the last section endpoint and the route endpoint aren't the same then the direct route is taken between these two with the default speed to reach the end.