Fixed mainmenu swipe visibly starting in wrong page.
Previously the mainmenu page starting from page 0 (dailytask) and was moved to page 2 (home) at the end of first frame start.
However due to the accidentally calling the CurrentPage, it skipped the instant page shift and instead used to normal scroll effect since it was started before the instant shift was called.