BattleCamera: Expanded view and Added controls
Added TargetSizeMin which defines the area that should be always visible
Added TargetSizeMax which defines the area that the camera can expand to when fitting the min area on screen (anything beyond the max area is filled with solid color)

Added SetView method that enables the camera and sets scale, offset and rotation. scale effects the scale of the min area. offset effects offset of the camera while ensuring that the min area stays on screen. rotate bool replaces the Rotate180 method.
Added UnsetView method that disables the camera

Removed the CameraAspectRatio component since it would interfere with BattleCamera

At the start the camera is in the unset state until SetView is called