GFX Update - Custom avatar assets & Head Masks
Updated / replaced older custom avatar textures.

Newly requested separation of the Left and the Right side of eyebrows, eyes, feet & hands. Naming scheme AvatarXXXPartL = Left side of character's part in question etc.

Avatar Head Masks. To be used with the custom avatars head texture, so that the body will not be rendered on top of the body. Includes black and white mask to cover different possible uses for the whole heads shape.