Changed the interface solution into a global singleton solution.
I hadn't checked that the Unity doesn't serialize interfaces, which effectively killed the rational behind the previous solution.
I'll change the references to be global singletons in the same way I have done elsewhere. This should allow access the the music list from the selection box when the correct enum is selected.