Added a handler for the reaction selector objects.
Data to the Object is now set throught the handler.
This includes the button listener to select the emotion.
This button now invokes a event delegate that the main reaction handler takes to update itself.
AddReaction method now takes a Mood enum and uses it to fetch the correct sprite and object.
Renamed the ChatAddReactions to MessageReactionsHandler.