Cleaned up the structure of PlayerAvatar
Cleaned up and improved the structure of PlayerAvatar to make it more robust against future implementations. Changed the featureIDs to be separate fields instead of a list and added a method to assign values based on the feature supplied. Change featurepicker to take PlayerAvatar straight instead of a list of parts. ColourPicker now uses the Color field. Edited the AvatarData creation in AvatarEditorController to loop over the features provided and assign them individually to AvatarData.