Moved the PollManager lists into a separate singleton script.
The fact that were static was causing issues with updating and fetching the current data. Sometimes it would place the whole poll list into the current poll list unnecessarily.
Also changed the load method to sort the poll into current and past polls.
Removed the redundant poll creation from when updating the lists.